+++ src/Model/Entity/User.php
@@ -13,6 +13,7 @@
 namespace MeCms\Model\Entity;
 
 use Cake\Auth\DefaultPasswordHasher;
+use Cake\Filesystem\Folder;
 use Cake\ORM\Entity;
 
 /**
@@ -50,7 +51,7 @@ class User extends Entity
      * Virtual fields that should be exposed
      * @var array
      */
-    protected $_virtual = ['full_name'];
+    protected $_virtual = ['full_name', 'picture'];
 
     /**
      * Gets the full name (virtual field)
@@ -59,6 +60,28 @@ class User extends Entity
     protected function _getFullName()
     {
         return sprintf('%s %s', $this->_properties['first_name'], $this->_properties['last_name']);
+    }
+
+    /**
+     * Gets the picture (virtual field)
+     * @return string
+     */
+    protected function _getPicture()
+    {
+        if (!empty($this->_properties['id'])) {
+            $files = ((new Folder(USER_PICTURES))->find($this->_properties['id'] . '\..+'));
+
+            if (!empty($files)) {
+                return 'users' . DS . array_values($files)[0];
+            }
+        }
+
+        //Checks for `webroot/img/no-avatar.jpg`
+        if (is_readable(WWW_ROOT . 'img' . DS . 'no-avatar.jpg')) {
+            return 'no-avatar.jpg';
+        }
+
+        return ME_CMS . '.no-avatar.jpg';
     }
 
     /**